+++ src/lib.rs
@@ -38,6 +38,8 @@ pub mod manifest;
 pub mod package;
 /// Supported registries
 pub mod registry;
+/// Resolve package dependencies.
+pub mod resolver;
 
 /// Cargo build integration for buffrs
 ///
